rpc: Move operator code to mod.rs.
